// techpalava-admin-portal/types.ts
export type SubmissionType = 'News' | 'Job' | 'Event' | 'Advertise' | 'Contact';
export type SubmissionStatus = 'New' | 'In Review' | 'Approved' | 'Rejected';

export interface Submission {
  id: string;
  type: SubmissionType;
  title: string;
  submittedBy: {
    name: string;
    email: string;
    company?: string;
  };
  content: string;
  dateSubmitted: string;
  status: SubmissionStatus;
  source: string;
  isRead: boolean;
  internalNotes?: string;
  attachments?: string[];
}